--- serverside/src/class/class.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ClassController } from './class.controller';
-
-describe('ClassController', () => {
-  let controller: ClassController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [ClassController],
-    }).compile();
-
-    controller = module.get<ClassController>(ClassController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
