+++ serverside/tests/class/class.controller.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UsersController } from './users.controller';
+import { ClassController } from './../../src/class/class.controller';
 
-describe('UsersController', () => {
-  let controller: UsersController;
+describe('ClassController', () => {
+  let controller: ClassController;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [UsersController],
+      controllers: [ClassController],
     }).compile();
 
-    controller = module.get<UsersController>(UsersController);
+    controller = module.get<ClassController>(ClassController);
   });
 
   it('should be defined', () => {
