+++ serverside/tests/event/event.controller.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UserController } from './user.controller';
+import { EventController } from './../../src/event/event.controller';
 
-describe('UserController', () => {
-  let controller: UserController;
+describe('EventController', () => {
+  let controller: EventController;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [UserController],
+      controllers: [EventController],
     }).compile();
 
-    controller = module.get<UserController>(UserController);
+    controller = module.get<EventController>(EventController);
   });
 
   it('should be defined', () => {
