+++ serverside/tests/event/event.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { ClassService } from './class.service';
+import { EventService } from './../../src/event/event.service';
 
-describe('ClassService', () => {
-  let service: ClassService;
+describe('EventService', () => {
+  let service: EventService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [ClassService],
+      providers: [EventService],
     }).compile();
 
-    service = module.get<ClassService>(ClassService);
+    service = module.get<EventService>(EventService);
   });
 
   it('should be defined', () => {
