+++ serverside/tests/student/student.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { StudentController } from './../../src/student/student.controller';
+
+describe('StudentController', () => {
+  let controller: StudentController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [StudentController],
+    }).compile();
+
+    controller = module.get<StudentController>(StudentController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
