+++ serverside/tests/student/student.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UserService } from './user.service';
+import { StudentService } from './../../src/student/student.service';
 
-describe('UserService', () => {
-  let service: UserService;
+describe('StudentService', () => {
+  let service: StudentService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [UserService],
+      providers: [StudentService],
     }).compile();
 
-    service = module.get<UserService>(UserService);
+    service = module.get<StudentService>(StudentService);
   });
 
   it('should be defined', () => {
