+++ serverside/tests/teacher/teacher.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { TeacherController } from './teacher.controller';
+import { TeacherController } from './../../src/teacher/teacher.controller';
 
 describe('TeacherController', () => {
   let controller: TeacherController;
