+++ serverside/tests/test/test.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UsersService } from './users.service';
+import { TestService } from './../../src/test/test.service';
 
-describe('UsersService', () => {
-  let service: UsersService;
+describe('TestService', () => {
+  let service: TestService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [UsersService],
+      providers: [TestService],
     }).compile();
 
-    service = module.get<UsersService>(UsersService);
+    service = module.get<TestService>(TestService);
   });
 
   it('should be defined', () => {
